Build Ceres and SuiteSparse using CMake and FetchContent (#10)
Basically a complete overhaul of the build system. The key change here is the removal of GPL licensed SuiteSparse components. We still need LGPL licensed components, but this is much better than before. This also enables the use of the Eigen we have in allwpilib, which will enable wpimath to be used in wpical.